redex delim-cont example: fix `if' in `get-marks'
The `if' case of `get-marks' was missing subscripts, which
caused it not to match and defer to an `if' in `get-marks-core',
but `get-marks-core' is not supposed to have an `if' rule.
Fix the one in `get-marks' and removed it from `get-marks-core'.

original commit: 39f2aa7f1ab72f39fc3f765ea886eb2358ce1bde