redex: fix stepper
This was exposed by the removal of the undefined value.

It's not clear if there was a bug lurking here or not.
The call to "move" was being bypassed on the initialization
of "pb", so the side effect of setting "set-visible?" wouldn't
have happened. Either that really didn't cause any problems,
or perhaps things looked strange briefly until the user did
something.

original commit: 8ed5a32d5db0aae43594ff249fcb28e1625abb02