add many parameters and hooks for rendering control
Expose several separation amounts as parameters, and add a hook
for pict adjustments at various line and rule levels.

Some rendering strategies based on `table` have been replaced with a
more direct layers-of-boxes approach. That apporach makes it easier to
normalize separation parameters and to make the pict-adjusting hook
work in general. The end result should be exactly as before when using
the default values for new parameters --- except that the font size
for rendering metafunction names in contracts has been corrected.