Adjust define-metafunction and reduction-relation so they generate less code
The code corresponding to the side-conditions and where clauses (etc) that
can appear in a define-metafunction and reduction-relation clause was being
generated twice -- once to actually run the metafunction and once with
a small variation that combined results to build a boolean instead of the actual
result (the boolean indicated if any of the where clauses or side-conditions
caused the corresponding case to fail).

The boolean version of the code was being used by the random generator
to tell if generation based on the left-hand side patterns would have
failed. This was somewhat useful, but the only thing the generator
did with the boolean was throw away the generated term and try to
generate another one when it was #f and the amount of code that was being
generated was non-trival. So this commit just drops that functionality.

The redex/examples/r6rs file dropped in size from 272k to 224k and the
time it takes to compile dropped from about 3.9 seconds to 3.2 seconds,
using the command:

$ racket -l racket/base -e '(time (dynamic-require (quote redex/examples/r6rs/r6rs) #f))'

The time it takes to compile with errortrace dropped roughly
proportionally, from 12.2 second to 8.8 seconds, using this command to
measure the time:

$ racket -l racket/base -l errortrace -e '(time (dynamic-require "r6rs.rkt" #f))

(in both cases, without the .zo file for r6rs.rkt)