Fix bind-withs so that it lifts out the calls to compile-language
This has the benefit of making judgment-form extensions work properly
when the language changes and it has a relatively narrow performance
improvement (as below).

It also has a backwards incompatibility. This program's behavior changed:

  #lang racket
  (require redex/reduction-semantics)
  (define-language grammar)
  (apply-reduction-relation
   (reduction-relation
    grammar
    (--> any
         1
         (where number_1 2)
         (where (side-condition any (number? (term number_1))) dontcare)))
    'dontcare)

Before this commit, `number_1` was bound, but after it, the expressions in
`side-condition` patterns now get only the variables in the surrounding context,
they don't see the earlier `where` clauses. (The right-hand sides of the `where`
clauses still see the earlier `where` clauses; this is only for expressions
inside `side-condition` that are inside patterns)

This change applies to patterns that are in `where` clauses in
reduction relations, metafunctions, and judgment forms, as well as
judgment-holds clauses in those definitions as well as premises in
judgment forms.

The performance improvement is probably mostly not noticeable because
of the agressive caching that Redex does, but the program below
defeats the caching and I'm seeing about a 1.8 performance improvement
for it.  The win comes from the fact that the `where` clause patterns
are no longer being compiled in the loop; they are compiled only once,
when the metafunction definition itself is evaluated

  #lang racket
  (require redex/reduction-semantics)
  (define-language L)
  (define-metafunction L
    [(m any)
     0
     (where (any_11 any_22 any_23 any_24 any_25 any_26 any_27) any)
     (where (any_11 any_22 any_23 any_24 any_25 any_26 any_27) any)
     (where (any_11 any_22 any_23 any_24 any_25 any_26 any_27) any)
     (where (any_11 any_22 any_23 any_24 any_25 any_26 any_27) any)
     (where (any_11 any_22 any_23 any_24 any_25 any_26 any_27) any)])
  (collect-garbage) (collect-garbage) (collect-garbage)
  (time
   (let ([l '(1 2 3 4 5 6)])
     (for ([x (in-range 10000)])
       (term (m ,(cons x l))))))