add 'mf-apply' keyword
In a term, `(mf-apply f arg ...)` asserts that `f` is a metafunction
 and expands to `(f arg ...)`.
Outside of a term, a `mf-apply` is an error.

Typesetting terms with `mf-apply` annotations makes the same picts/bitmaps
as terms without the annotations, but the intermediate `lw` struct for
a term using `mf-apply` currently has a wider column span.