add term-fn-id? predicate
`(term-fn-id? id)` takes the `syntax-local-value` of `id` and checks that
it is a `term-fn?`

Using the predicate simplifies a few lines in `term.rkt`

It also fixes an error in `define-metafunction/extension`, which wasn't
checking whether the existing metafunction was a `term-fn?`.

For example, this program:
```
  #lang racket
  (require redex)

  (define-syntax (fake-mf stx)
    (syntax (term 1)))

  (define-language nats (nat Z (S Z)))

  (define-metafunction/extension fake-mf nats
    ((g) ()))
```

Used to raise an error like:
```
  term-fn-field0: contract violation
    expected: term-fn?
```