improve an internal data structure
related to #135 -- the use of `second` in the error in the example is
because lists of length two were being used instead of a struct that
contains a redex term and a binding map. This commit changes that to
use a struct. The error persists, however, but at least it happens
earlier now.